Moved all lib index.ts to lib roots and fixed public_api.js imports
