+++ libs/angular-core/index.ts
