+++ libs/angular-styles/index.scss
@@ -1,1 +1,1 @@
-@import 'src/index.scss';
+@import 'src/lib/index.scss';
