+++ libs/angular-ui/public_api.ts
@@ -1,3 +1,3 @@
-export { UiModule } from './src/ui.module'
-export { UiLink, LayoutComponent, LayoutBaseComponent, UiButton } from './src/modules'
-export { UiService, UiProfile, UiNotification, UiLayout, UiBadge} from './src/services/ui.service'
+export { UiModule } from './src/lib/ui.module'
+export { UiLink, LayoutComponent, LayoutBaseComponent, UiButton } from './src/lib/modules'
+export { UiService, UiProfile, UiNotification, UiLayout, UiBadge} from './src/lib/services/ui.service'
